refactor: modularize code crafter tests
