+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -162,6 +162,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Brought down from 892 to 148 lines by modularizing handshake, routing, authentication, and logging utilities. Protocol file now orchestrates these modules.
 
+### CRIT-017: Code Crafter Test Suite ✅
+- **File**: `tests/ai_ml/test_code_crafter.py`
+- **Status**: Completed
+- **Assigned To**: Victor Dixon
+- **Completion Date**: 2025-08-24
+- **Summary**: Refactored monolithic 976-line test file into modular models, generation, validation, and integration suites with a lightweight orchestrator.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
