+++ tests/ai_ml/test_code_crafter.py
@@ -1,976 +1,29 @@
-"""
-🧪 CodeCrafter Tests
-AI & ML Integration Specialist - TDD Integration Project
+"""Orchestrator for the modular CodeCrafter test suite.
 
-Test-Driven Development implementation for CodeCrafter module
-"""
+This file re-exports the main test classes so legacy imports of
+``tests.ai_ml.test_code_crafter`` continue to function. Each test class
+is defined in its own module focusing on a single concern:
 
-import os
-import json
-import tempfile
-import logging
-
-from src.utils.stability_improvements import stability_manager, safe_import
-from pathlib import Path
-from typing import Dict, Optional, Any, List
-from dataclasses import dataclass
-from unittest.mock import Mock, patch, mock_open
-import pytest
-
-# Configure logging
-logger = logging.getLogger(__name__)
-
-# ============================================================================
-# EMBEDDED MODULE CODE (to avoid import issues)
-# ============================================================================
-
-
-@dataclass
-class CodeAnalysis:
-    """Results of code analysis"""
-
-    file_path: str
-    complexity_score: float
-    maintainability_index: float
-    code_smells: List[str] = None
-    suggestions: List[str] = None
-    security_issues: List[str] = None
-    performance_issues: List[str] = None
-    documentation_coverage: float = 0.0
-    test_coverage: float = 0.0
-
-    def __post_init__(self):
-        if self.code_smells is None:
-            self.code_smells = []
-        if self.suggestions is None:
-            self.suggestions = []
-        if self.security_issues is None:
-            self.security_issues = []
-        if self.performance_issues is None:
-            self.performance_issues = []
-
-
-@dataclass
-class CodeGenerationRequest:
-    """Request for code generation"""
-
-    description: str
-    language: str
-    framework: Optional[str] = None
-    requirements: List[str] = None
-    constraints: List[str] = None
-    style_guide: Optional[str] = None
-    include_tests: bool = True
-    include_docs: bool = True
-
-    def __post_init__(self):
-        if self.requirements is None:
-            self.requirements = []
-        if self.constraints is None:
-            self.constraints = []
-
-
-@dataclass
-class CodeGenerationResult:
-    """Result of code generation"""
-
-    code: str
-    explanation: str
-    estimated_complexity: float
-    tests: Optional[str] = None
-    documentation: Optional[str] = None
-    dependencies: List[str] = None
-    usage_examples: List[str] = None
-
-    def __post_init__(self):
-        if self.dependencies is None:
-            self.dependencies = []
-        if self.usage_examples is None:
-            self.usage_examples = []
-
-
-class CodeCrafter:
-    """AI-powered code generation and analysis tool"""
-
-    def __init__(self):
-        self.supported_languages = [
-            "python",
-            "javascript",
-            "typescript",
-            "java",
-            "cpp",
-            "csharp",
-            "go",
-            "rust",
-        ]
-        self.supported_frameworks = {
-            "python": ["django", "flask", "fastapi", "pytorch", "tensorflow", "pandas"],
-            "javascript": ["react", "vue", "angular", "node", "express"],
-            "typescript": ["react", "vue", "angular", "node", "express"],
-            "java": ["spring", "hibernate", "junit", "maven"],
-            "cpp": ["boost", "qt", "opencv", "eigen"],
-            "csharp": [".net", "asp.net", "entity", "xunit"],
-            "go": ["gin", "echo", "gorm", "testify"],
-            "rust": ["tokio", "serde", "actix", "criterion"],
-        }
-
-    def generate_code(self, request: CodeGenerationRequest) -> CodeGenerationResult:
-        """
-        Generate code based on natural language description
-
-        Args:
-            request: Code generation request
-
-        Returns:
-            Generated code with tests and documentation
-        """
-        logger.info(f"Generating {request.language} code for: {request.description}")
-
-        try:
-            # Validate request
-            self._validate_generation_request(request)
-
-            # Generate code using AI (mocked for testing)
-            code = self._generate_code_content(request)
-            tests = self._generate_tests(request) if request.include_tests else None
-            documentation = (
-                self._generate_documentation(request) if request.include_docs else None
-            )
-
-            # Calculate complexity
-            complexity = self._calculate_complexity(code)
-
-            # Extract dependencies
-            dependencies = self._extract_dependencies(code, request.language)
-
-            # Generate usage examples
-            examples = self._generate_usage_examples(code, request.language)
-
-            return CodeGenerationResult(
-                code=code,
-                tests=tests,
-                documentation=documentation,
-                explanation=f"Generated {request.language} code for: {request.description}",
-                estimated_complexity=complexity,
-                dependencies=dependencies,
-                usage_examples=examples,
-            )
-
-        except Exception as e:
-            logger.error(f"Code generation failed: {e}")
-            raise
-
-    def analyze_code(self, file_path: str) -> CodeAnalysis:
-        """
-        Analyze existing code for quality metrics
-
-        Args:
-            file_path: Path to the code file
-
-        Returns:
-            Code analysis results
-        """
-        logger.info(f"Analyzing code file: {file_path}")
-
-        try:
-            # Read and parse code
-            code_content = self._read_code_file(file_path)
-
-            # Analyze complexity
-            complexity = self._analyze_complexity(code_content)
-
-            # Analyze maintainability
-            maintainability = self._analyze_maintainability(code_content)
-
-            # Detect code smells
-            smells = self._detect_code_smells(code_content)
-
-            # Generate suggestions
-            suggestions = self._generate_suggestions(code_content, smells)
-
-            # Check for security issues
-            security_issues = self._check_security_issues(code_content)
-
-            # Check for performance issues
-            performance_issues = self._check_performance_issues(code_content)
-
-            # Calculate documentation coverage
-            doc_coverage = self._calculate_documentation_coverage(code_content)
-
-            # Calculate test coverage (would need test files)
-            test_coverage = 0.0
-
-            return CodeAnalysis(
-                file_path=file_path,
-                complexity_score=complexity,
-                maintainability_index=maintainability,
-                code_smells=smells,
-                suggestions=suggestions,
-                security_issues=security_issues,
-                performance_issues=performance_issues,
-                documentation_coverage=doc_coverage,
-                test_coverage=test_coverage,
-            )
-
-        except Exception as e:
-            logger.error(f"Code analysis failed: {e}")
-            raise
-
-    def _validate_generation_request(self, request: CodeGenerationRequest) -> None:
-        """Validate code generation request"""
-        if not request.description:
-            raise ValueError("Description is required")
-
-        if request.language not in self.supported_languages:
-            raise ValueError(f"Unsupported language: {request.language}")
-
-        if (
-            request.framework
-            and request.framework
-            not in self.supported_frameworks.get(request.language, [])
-        ):
-            raise ValueError(
-                f"Unsupported framework: {request.framework} for language: {request.language}"
-            )
-
-    def _generate_code_content(self, request: CodeGenerationRequest) -> str:
-        """Generate the main code content"""
-        # Mock implementation for testing
-        if request.language == "python":
-            return self._generate_python_code(request)
-        elif request.language == "javascript":
-            return self._generate_javascript_code(request)
-        else:
-            return f"// Generated {request.language} code for: {request.description}"
-
-    def _generate_python_code(self, request: CodeGenerationRequest) -> str:
-        """Generate Python code"""
-        code = f'''"""
-{request.description}
-Generated by CodeCrafter
+* ``test_code_crafter_models`` – dataclass validations
+* ``test_code_crafter_generation`` – code generation workflow
+* ``test_code_crafter_validation`` – analysis and helper methods
+* ``test_code_crafter_integration`` – end-to-end scenarios
 """
 
-{self._generate_python_imports(request)}
-
-def main():
-    """Main function"""
-    print("Hello from generated code!")
-
-    # TODO: Implement your logic here
-    pass
-
-if __name__ == "__main__":
-    main()
-'''
-        return code
-
-    def _generate_javascript_code(self, request: CodeGenerationRequest) -> str:
-        """Generate JavaScript code"""
-        code = f"""/**
- * {request.description}
- * Generated by CodeCrafter
- */
-
-{self._generate_javascript_imports(request)}
-
-function main() {{
-    console.log("Hello from generated code!");
-
-    // TODO: Implement your logic here
-}}
-
-// Export for module usage
-if (typeof module !== 'undefined' && module.exports) {{
-    module.exports = {{ main }};
-}}
-"""
-        return code
-
-    def _generate_python_imports(self, request: CodeGenerationRequest) -> str:
-        """Generate Python imports based on framework"""
-        if request.framework == "flask":
-            return "from flask import Flask, request, jsonify\n\napp = Flask(__name__)"
-        elif request.framework == "pandas":
-            return "import pandas as pd\nimport numpy as np"
-        else:
-            return "import os\nimport sys"
-
-    def _generate_javascript_imports(self, request: CodeGenerationRequest) -> str:
-        """Generate JavaScript imports based on framework"""
-        if request.framework == "express":
-            return "const express = require('express');\nconst app = express();"
-        elif request.framework == "react":
-            return "import React from 'react';"
-        else:
-            return "// Standard JavaScript"
-
-    def _generate_tests(self, request: CodeGenerationRequest) -> str:
-        """Generate test code"""
-        if request.language == "python":
-            return self._generate_python_tests(request)
-        elif request.language == "javascript":
-            return self._generate_javascript_tests(request)
-        else:
-            return f"// Tests for {request.language}"
-
-    def _generate_python_tests(self, request: CodeGenerationRequest) -> str:
-        """Generate Python tests"""
-        return f'''"""
-Tests for generated code
-"""
-
-import unittest
-from unittest.mock import patch
-
-class TestGeneratedCode(unittest.TestCase):
-    """Test cases for generated code"""
-
-    def test_main_function(self):
-        """Test main function exists"""
-        # TODO: Add actual test logic
-        self.assertTrue(True)
-
-if __name__ == '__main__':
-    unittest.main()
-'''
-
-    def _generate_javascript_tests(self, request: CodeGenerationRequest) -> str:
-        """Generate JavaScript tests"""
-        return f"""/**
- * Tests for generated code
- */
-
-// TODO: Add actual test framework setup
-describe('Generated Code', () => {{
-    it('should have main function', () => {{
-        // TODO: Add actual test logic
-        expect(true).toBe(true);
-    }});
-}});
-"""
-
-    def _generate_documentation(self, request: CodeGenerationRequest) -> str:
-        """Generate documentation"""
-        # Capitalize language name for display with special cases
-        language_display = request.language.capitalize()
-        if request.language == "javascript":
-            language_display = "JavaScript"
-        elif request.language == "typescript":
-            language_display = "TypeScript"
-
-        return f"""# {request.description}
-
-## Overview
-This code was generated by CodeCrafter for the following requirements:
-- Language: {language_display}
-- Framework: {request.framework or 'None'}
-- Requirements: {', '.join(request.requirements) if request.requirements else 'None'}
-
-## Usage
-```{request.language}
-# Example usage code here
-```
-
-## Dependencies
-- TODO: List dependencies
-
-## Testing
-Run the tests using the appropriate test runner for {language_display}.
-"""
-
-    def _calculate_complexity(self, code: str) -> float:
-        """Calculate code complexity"""
-        # Simple complexity calculation based on lines and keywords
-        lines = code.split("\n")
-        complexity = len(lines) * 0.1
-
-        # Add complexity for certain keywords
-        complexity_keywords = ["if", "for", "while", "try", "except", "catch"]
-        for keyword in complexity_keywords:
-            complexity += code.count(keyword) * 0.5
-
-        return round(complexity, 2)
-
-    def _extract_dependencies(self, code: str, language: str) -> List[str]:
-        """Extract dependencies from code"""
-        dependencies = []
-
-        if language == "python":
-            import_lines = [
-                line
-                for line in code.split("\n")
-                if line.strip().startswith("import") or line.strip().startswith("from")
-            ]
-            for line in import_lines:
-                if "import" in line:
-                    # Handle "import module" and "import module as alias"
-                    if line.strip().startswith("import"):
-                        parts = line.split("import")[1].strip()
-                        if " as " in parts:
-                            module = parts.split(" as ")[0].strip()
-                        else:
-                            module = parts.strip()
-                        dependencies.append(module)
-                    elif line.strip().startswith("from"):
-                        # Handle "from module import item"
-                        parts = line.split("from")[1].split("import")[0].strip()
-                        dependencies.append(parts)
-
-        elif language == "javascript":
-            import_lines = [
-                line
-                for line in code.split("\n")
-                if "require(" in line or "import" in line
-            ]
-            for line in import_lines:
-                if "require(" in line:
-                    dep = line.split("require(")[1].split(")")[0].strip("'\"")
-                    dependencies.append(dep)
-
-        return dependencies
-
-    def _generate_usage_examples(self, code: str, language: str) -> List[str]:
-        """Generate usage examples"""
-        examples = []
-
-        if language == "python":
-            examples.append("python main.py")
-            examples.append("from generated_code import main")
-            examples.append("main()")
-
-        elif language == "javascript":
-            examples.append("node main.js")
-            examples.append("const { main } = require('./main.js')")
-            examples.append("main()")
-
-        return examples
-
-    def _read_code_file(self, file_path: str) -> str:
-        """Read code file content"""
-        with open(file_path, "r", encoding="utf-8") as f:
-            return f.read()
-
-    def _analyze_complexity(self, code: str) -> float:
-        """Analyze code complexity"""
-        return self._calculate_complexity(code)
-
-    def _analyze_maintainability(self, code: str) -> float:
-        """Analyze code maintainability"""
-        # Simple maintainability score (0-100)
-        lines = code.split("\n")
-        score = 100.0
-
-        # Reduce score for long functions
-        if len(lines) > 50:
-            score -= 20
-
-        # Reduce score for complex logic
-        complexity = self._analyze_complexity(code)
-        if complexity > 10:
-            score -= 30
-
-        return max(0.0, score)
-
-    def _detect_code_smells(self, code: str) -> List[str]:
-        """Detect code smells"""
-        smells = []
-
-        # Check for long functions
-        lines = code.split("\n")
-        if len(lines) > 50:
-            smells.append("Long function detected (>50 lines)")
-
-        # Check for magic numbers
-        import re
-
-        magic_numbers = re.findall(r"\b\d{3,}\b", code)
-        if magic_numbers:
-            smells.append(f"Magic numbers detected: {', '.join(set(magic_numbers))}")
-
-        # Check for commented code
-        if "#" in code and "TODO" in code:
-            smells.append("Commented code or TODOs detected")
-
-        return smells
-
-    def _generate_suggestions(self, code: str, smells: List[str]) -> List[str]:
-        """Generate improvement suggestions"""
-        suggestions = []
-
-        if "Long function detected" in str(smells):
-            suggestions.append(
-                "Consider breaking down long functions into smaller, focused functions"
-            )
-
-        if "Magic numbers detected" in str(smells):
-            suggestions.append("Replace magic numbers with named constants")
-
-        if "Commented code or TODOs detected" in str(smells):
-            suggestions.append("Remove commented code and implement TODO items")
-
-        return suggestions
-
-    def _check_security_issues(self, code: str) -> List[str]:
-        """Check for security issues"""
-        issues = []
-
-        # Check for hardcoded secrets
-        if "password" in code.lower() or "secret" in code.lower():
-            issues.append("Potential hardcoded secrets detected")
-
-        # Check for SQL injection patterns
-        if "execute(" in code.lower() and "sql" in code.lower():
-            issues.append("Potential SQL injection vulnerability")
-
-        return issues
-
-    def _check_performance_issues(self, code: str) -> List[str]:
-        """Check for performance issues"""
-        issues = []
-
-        # Check for nested loops
-        if code.count("for") > 1 or code.count("while") > 1:
-            issues.append("Nested loops detected - consider optimization")
-
-        # Check for large data structures
-        if "list(" in code or "dict(" in code:
-            issues.append("Large data structure creation detected")
-
-        return issues
-
-    def _calculate_documentation_coverage(self, code: str) -> float:
-        """Calculate documentation coverage"""
-        lines = code.split("\n")
-        doc_lines = 0
-
-        for line in lines:
-            stripped = line.strip()
-            if (
-                stripped.startswith('"""')
-                or stripped.startswith("'''")
-                or stripped.startswith("#")
-            ):
-                doc_lines += 1
-
-        if len(lines) == 0:
-            return 0.0
-
-        return round((doc_lines / len(lines)) * 100, 2)
-
-
-# ============================================================================
-# TEST IMPLEMENTATION
-# ============================================================================
-
-
-class TestCodeAnalysis:
-    """Test CodeAnalysis dataclass."""
-
-    @pytest.mark.unit
-    def test_code_analysis_creation(self):
-        """Test creating CodeAnalysis with all parameters."""
-        analysis = CodeAnalysis(
-            file_path="test.py",
-            complexity_score=5.5,
-            maintainability_index=85.0,
-            code_smells=["Long function"],
-            suggestions=["Break into smaller functions"],
-            security_issues=["Hardcoded password"],
-            performance_issues=["Nested loops"],
-            documentation_coverage=75.0,
-            test_coverage=60.0,
-        )
-
-        assert analysis.file_path == "test.py"
-        assert analysis.complexity_score == 5.5
-        assert analysis.maintainability_index == 85.0
-        assert "Long function" in analysis.code_smells
-        assert "Break into smaller functions" in analysis.suggestions
-        assert "Hardcoded password" in analysis.security_issues
-        assert "Nested loops" in analysis.performance_issues
-        assert analysis.documentation_coverage == 75.0
-        assert analysis.test_coverage == 60.0
-
-    @pytest.mark.unit
-    def test_code_analysis_defaults(self):
-        """Test creating CodeAnalysis with minimal parameters."""
-        analysis = CodeAnalysis(
-            file_path="test.py", complexity_score=3.0, maintainability_index=90.0
-        )
-
-        assert analysis.file_path == "test.py"
-        assert analysis.complexity_score == 3.0
-        assert analysis.maintainability_index == 90.0
-        assert analysis.code_smells == []
-        assert analysis.suggestions == []
-        assert analysis.security_issues == []
-        assert analysis.performance_issues == []
-        assert analysis.documentation_coverage == 0.0
-        assert analysis.test_coverage == 0.0
-
-
-class TestCodeGenerationRequest:
-    """Test CodeGenerationRequest dataclass."""
-
-    @pytest.mark.unit
-    def test_code_generation_request_creation(self):
-        """Test creating CodeGenerationRequest with all parameters."""
-        request = CodeGenerationRequest(
-            description="Create a web API endpoint",
-            language="python",
-            framework="flask",
-            requirements=["RESTful", "JSON response"],
-            constraints=["No external dependencies"],
-            style_guide="PEP 8",
-            include_tests=True,
-            include_docs=True,
-        )
-
-        assert request.description == "Create a web API endpoint"
-        assert request.language == "python"
-        assert request.framework == "flask"
-        assert "RESTful" in request.requirements
-        assert "No external dependencies" in request.constraints
-        assert request.style_guide == "PEP 8"
-        assert request.include_tests is True
-        assert request.include_docs is True
-
-    @pytest.mark.unit
-    def test_code_generation_request_minimal(self):
-        """Test creating CodeGenerationRequest with minimal parameters."""
-        request = CodeGenerationRequest(
-            description="Simple function", language="javascript"
-        )
-
-        assert request.description == "Simple function"
-        assert request.language == "javascript"
-        assert request.framework is None
-        assert request.requirements == []
-        assert request.constraints == []
-        assert request.style_guide is None
-        assert request.include_tests is True
-        assert request.include_docs is True
-
-
-class TestCodeGenerationResult:
-    """Test CodeGenerationResult dataclass."""
-
-    @pytest.mark.unit
-    def test_code_generation_result_creation(self):
-        """Test creating CodeGenerationResult with all parameters."""
-        result = CodeGenerationResult(
-            code="def hello(): print('Hello')",
-            tests="def test_hello(): assert True",
-            documentation="# Hello function",
-            explanation="Simple greeting function",
-            estimated_complexity=1.0,
-            dependencies=["builtins"],
-            usage_examples=["hello()"],
-        )
-
-        assert result.code == "def hello(): print('Hello')"
-        assert result.tests == "def test_hello(): assert True"
-        assert result.documentation == "# Hello function"
-        assert result.explanation == "Simple greeting function"
-        assert result.estimated_complexity == 1.0
-        assert "builtins" in result.dependencies
-        assert "hello()" in result.usage_examples
-
-    @pytest.mark.unit
-    def test_code_generation_result_minimal(self):
-        """Test creating CodeGenerationResult with minimal parameters."""
-        result = CodeGenerationResult(
-            code="print('Hello')",
-            explanation="Simple print statement",
-            estimated_complexity=0.5,
-        )
-
-        assert result.code == "print('Hello')"
-        assert result.tests is None
-        assert result.documentation is None
-        assert result.explanation == "Simple print statement"
-        assert result.estimated_complexity == 0.5
-        assert result.dependencies == []
-        assert result.usage_examples == []
-
-
-class TestCodeCrafter:
-    """Test CodeCrafter class."""
-
-    @pytest.fixture
-    def temp_code_dir(self):
-        """Create temporary directory for code files."""
-        temp_dir = tempfile.mkdtemp()
-        yield Path(temp_dir)
-        import shutil
-
-        shutil.rmtree(temp_dir)
-
-    @pytest.fixture
-    def sample_python_code(self):
-        """Sample Python code for testing."""
-        return '''"""
-Sample Python code for testing
-"""
-
-def fibonacci(n):
-    """Calculate fibonacci number."""
-    if n <= 1:
-        return n
-    return fibonacci(n-1) + fibonacci(n-2)
-
-def main():
-    """Main function."""
-    result = fibonacci(10)
-    print(f"Fibonacci(10) = {result}")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    @pytest.mark.unit
-    def test_code_crafter_initialization(self):
-        """Test CodeCrafter initialization."""
-        crafter = CodeCrafter()
-
-        assert "python" in crafter.supported_languages
-        assert "javascript" in crafter.supported_languages
-        assert "flask" in crafter.supported_frameworks["python"]
-        assert "react" in crafter.supported_frameworks["javascript"]
-
-    @pytest.mark.unit
-    def test_generate_code_python(self):
-        """Test generating Python code."""
-        crafter = CodeCrafter()
-        request = CodeGenerationRequest(
-            description="Simple calculator function",
-            language="python",
-            include_tests=True,
-            include_docs=True,
-        )
-
-        result = crafter.generate_code(request)
-
-        assert result.code is not None
-        assert "def main()" in result.code
-        assert result.tests is not None
-        assert "unittest" in result.tests
-        assert result.documentation is not None
-        assert "Python" in result.documentation
-        assert result.explanation is not None
-        assert result.estimated_complexity > 0
-        assert len(result.dependencies) >= 0
-        assert len(result.usage_examples) > 0
-
-    @pytest.mark.unit
-    def test_generate_code_javascript(self):
-        """Test generating JavaScript code."""
-        crafter = CodeCrafter()
-        request = CodeGenerationRequest(
-            description="Web server",
-            language="javascript",
-            framework="express",
-            include_tests=True,
-            include_docs=True,
-        )
-
-        result = crafter.generate_code(request)
-
-        assert result.code is not None
-        assert "function main()" in result.code
-        assert result.tests is not None
-        assert "describe" in result.tests
-        assert result.documentation is not None
-        assert "JavaScript" in result.documentation
-        assert result.explanation is not None
-        assert result.estimated_complexity > 0
-        assert len(result.dependencies) >= 0
-        assert len(result.usage_examples) > 0
-
-    @pytest.mark.unit
-    def test_generate_code_invalid_language(self):
-        """Test generating code with invalid language."""
-        crafter = CodeCrafter()
-        request = CodeGenerationRequest(description="Test", language="invalid_language")
-
-        with pytest.raises(ValueError, match="Unsupported language"):
-            crafter.generate_code(request)
-
-    @pytest.mark.unit
-    def test_generate_code_invalid_framework(self):
-        """Test generating code with invalid framework."""
-        crafter = CodeCrafter()
-        request = CodeGenerationRequest(
-            description="Test", language="python", framework="invalid_framework"
-        )
-
-        with pytest.raises(ValueError, match="Unsupported framework"):
-            crafter.generate_code(request)
-
-    @pytest.mark.unit
-    def test_generate_code_no_description(self):
-        """Test generating code without description."""
-        crafter = CodeCrafter()
-        request = CodeGenerationRequest(description="", language="python")
-
-        with pytest.raises(ValueError, match="Description is required"):
-            crafter.generate_code(request)
-
-    @pytest.mark.unit
-    def test_analyze_code(self, temp_code_dir, sample_python_code):
-        """Test analyzing existing code."""
-        # Create test code file
-        code_file = temp_code_dir / "test_code.py"
-        with open(code_file, "w") as f:
-            f.write(sample_python_code)
-
-        crafter = CodeCrafter()
-        analysis = crafter.analyze_code(str(code_file))
-
-        assert analysis.file_path == str(code_file)
-        assert analysis.complexity_score > 0
-        assert analysis.maintainability_index > 0
-        assert isinstance(analysis.code_smells, list)
-        assert isinstance(analysis.suggestions, list)
-        assert isinstance(analysis.security_issues, list)
-        assert isinstance(analysis.performance_issues, list)
-        assert analysis.documentation_coverage >= 0
-        assert analysis.test_coverage == 0.0
-
-    @pytest.mark.unit
-    def test_analyze_code_file_not_found(self):
-        """Test analyzing non-existent code file."""
-        crafter = CodeCrafter()
-
-        with pytest.raises(FileNotFoundError):
-            crafter.analyze_code("nonexistent_file.py")
-
-    @pytest.mark.unit
-    def test_calculate_complexity(self):
-        """Test complexity calculation."""
-        crafter = CodeCrafter()
-        code = (
-            "def test():\n    if True:\n        for i in range(10):\n            pass"
-        )
-
-        complexity = crafter._calculate_complexity(code)
-
-        assert complexity > 0
-        assert isinstance(complexity, float)
-
-    @pytest.mark.unit
-    def test_extract_dependencies_python(self):
-        """Test dependency extraction from Python code."""
-        crafter = CodeCrafter()
-        code = "import os\nimport sys\nfrom flask import Flask"
-
-        dependencies = crafter._extract_dependencies(code, "python")
-
-        assert "os" in dependencies
-        assert "sys" in dependencies
-        assert "flask" in dependencies
-
-    @pytest.mark.unit
-    def test_extract_dependencies_javascript(self):
-        """Test dependency extraction from JavaScript code."""
-        crafter = CodeCrafter()
-        code = "const express = require('express');\nconst fs = require('fs');"
-
-        dependencies = crafter._extract_dependencies(code, "javascript")
-
-        assert "express" in dependencies
-        assert "fs" in dependencies
-
-    @pytest.mark.unit
-    def test_detect_code_smells(self):
-        """Test code smell detection."""
-        crafter = CodeCrafter()
-        code = "def very_long_function():\n" + "\n".join(["    pass"] * 60)
-
-        smells = crafter._detect_code_smells(code)
-
-        assert "Long function detected" in str(smells)
-
-    @pytest.mark.unit
-    def test_check_security_issues(self):
-        """Test security issue detection."""
-        crafter = CodeCrafter()
-        code = "password = 'secret123'"
-
-        issues = crafter._check_security_issues(code)
-
-        assert "hardcoded secrets" in str(issues).lower()
-
-    @pytest.mark.unit
-    def test_check_performance_issues(self):
-        """Test performance issue detection."""
-        crafter = CodeCrafter()
-        code = "for i in range(100):\n    for j in range(100):\n        pass"
-
-        issues = crafter._check_performance_issues(code)
-
-        assert "Nested loops" in str(issues)
-
-    @pytest.mark.unit
-    def test_calculate_documentation_coverage(self):
-        """Test documentation coverage calculation."""
-        crafter = CodeCrafter()
-        code = '"""Docstring"""\ndef test():\n    # Comment\n    pass'
-
-        coverage = crafter._calculate_documentation_coverage(code)
-
-        assert coverage > 0
-        assert coverage <= 100
-
-
-class TestCodeCrafterIntegration:
-    """Integration tests for CodeCrafter."""
-
-    @pytest.fixture
-    def temp_code_dir(self):
-        """Create temporary directory for code files."""
-        temp_dir = tempfile.mkdtemp()
-        yield Path(temp_dir)
-        import shutil
-
-        shutil.rmtree(temp_dir)
-
-    @pytest.mark.integration
-    def test_full_code_generation_workflow(self, temp_code_dir):
-        """Test complete code generation workflow."""
-        crafter = CodeCrafter()
-
-        # 1. Generate code
-        request = CodeGenerationRequest(
-            description="Data processing pipeline",
-            language="python",
-            framework="pandas",
-            requirements=["CSV input", "Data validation", "JSON output"],
-            include_tests=True,
-            include_docs=True,
-        )
-
-        result = crafter.generate_code(request)
-
-        # 2. Verify generation
-        assert result.code is not None
-        assert result.tests is not None
-        assert result.documentation is not None
-        assert "pandas" in result.dependencies
-
-        # 3. Save generated code
-        code_file = temp_code_dir / "generated_code.py"
-        with open(code_file, "w") as f:
-            f.write(result.code)
-
-        # 4. Analyze generated code
-        analysis = crafter.analyze_code(str(code_file))
-
-        # 5. Verify analysis
-        assert analysis.file_path == str(code_file)
-        assert analysis.complexity_score > 0
-        assert analysis.maintainability_index > 0
-
-        # 6. Check for generated files
-        assert code_file.exists()
-        assert code_file.stat().st_size > 0
+from .code_crafter_support import (
+    CodeAnalysis,
+    CodeGenerationRequest,
+    CodeGenerationResult,
+    CodeCrafter,
+)
+from .test_code_crafter_models import *  # noqa: F401,F403
+from .test_code_crafter_generation import *  # noqa: F401,F403
+from .test_code_crafter_validation import *  # noqa: F401,F403
+from .test_code_crafter_integration import *  # noqa: F401,F403
+
+__all__ = [
+    "CodeAnalysis",
+    "CodeGenerationRequest",
+    "CodeGenerationResult",
+    "CodeCrafter",
+]
